Added the file decoder
char, num1, num2, [num3], [num4]
First row must be the specification of the dimensions.
char must be d, num1 must be the number of blocks in x direction, num2 in y direction. num3&num4 are additional, those define the window size.
If you want to add the player char must be p, num1 is the x coord, num2 is the y coord.
If you want to add a block, char must be b, num1 is the x coord, num2 is the y coord, num3 is the block-ID.